+++ main/java/uk/co/real_logic/sbe/generation/TargetCodeGenerator.java
@@ -16,6 +16,7 @@
 package uk.co.real_logic.sbe.generation;
 
 import uk.co.real_logic.agrona.concurrent.UnsafeBuffer;
+import uk.co.real_logic.agrona.generation.PackageOutputManager;
 import uk.co.real_logic.sbe.SbeTool;
 import uk.co.real_logic.sbe.generation.cpp98.Cpp98Generator;
 import uk.co.real_logic.sbe.generation.cpp98.NamespaceOutputManager;
@@ -23,7 +24,6 @@ import uk.co.real_logic.sbe.generation.csharp.CSharpGenerator;
 import uk.co.real_logic.sbe.generation.csharp.CSharpNamespaceOutputManager;
 import uk.co.real_logic.sbe.generation.java.JavaGenerator;
 import uk.co.real_logic.sbe.generation.java.JavaMockPojoGenerator;
-import uk.co.real_logic.sbe.generation.java.PackageOutputManager;
 import uk.co.real_logic.sbe.generation.python.ModuleOutputManager;
 import uk.co.real_logic.sbe.generation.python.PythonGenerator;
 import uk.co.real_logic.sbe.ir.Ir;
