+++ main/java/uk/co/real_logic/sbe/ir/Encoding.java
@@ -17,7 +17,7 @@ package uk.co.real_logic.sbe.ir;
 
 import uk.co.real_logic.sbe.PrimitiveType;
 import uk.co.real_logic.sbe.PrimitiveValue;
-import uk.co.real_logic.sbe.util.Verify;
+import uk.co.real_logic.agrona.Verify;
 
 import java.nio.ByteOrder;
