+++ main/java/uk/co/real_logic/sbe/ir/HeaderStructure.java
@@ -16,7 +16,7 @@
 package uk.co.real_logic.sbe.ir;
 
 import uk.co.real_logic.sbe.PrimitiveType;
-import uk.co.real_logic.sbe.util.Verify;
+import uk.co.real_logic.agrona.Verify;
 
 import java.util.List;
