+++ main/java/uk/co/real_logic/sbe/ir/Ir.java
@@ -15,7 +15,7 @@
  */
 package uk.co.real_logic.sbe.ir;
 
-import uk.co.real_logic.sbe.util.Verify;
+import uk.co.real_logic.agrona.Verify;
 
 import java.util.*;
