+++ main/java/uk/co/real_logic/sbe/xml/MessageSchema.java
@@ -17,7 +17,7 @@
 package uk.co.real_logic.sbe.xml;
 
 import org.w3c.dom.Node;
-import uk.co.real_logic.sbe.util.Verify;
+import uk.co.real_logic.agrona.Verify;
 
 import java.nio.ByteOrder;
 import java.util.Collection;
