+++ test/java/uk/co/real_logic/sbe/generation/java/JavaGeneratorTest.java
@@ -20,8 +20,8 @@ import org.junit.Test;
 import uk.co.real_logic.agrona.MutableDirectBuffer;
 import uk.co.real_logic.agrona.concurrent.UnsafeBuffer;
 import uk.co.real_logic.agrona.generation.CompilerUtil;
+import uk.co.real_logic.agrona.generation.StringWriterOutputManager;
 import uk.co.real_logic.sbe.TestUtil;
-import uk.co.real_logic.sbe.generation.java.util.StringWriterOutputManager;
 import uk.co.real_logic.sbe.ir.Ir;
 import uk.co.real_logic.sbe.xml.IrGenerator;
 import uk.co.real_logic.sbe.xml.MessageSchema;
